persist auth and isAuthenticated, blacklist the others in authState. Move store logic to index.js
